New handler and states module
