+++ utils/states.py
@@ -1,0 +1,10 @@
+(
+    PROCEED,
+    FOLLOW_TELEGRAM,
+    FOLLOW_TWITTER,
+    SUBMIT_ADDRESS,
+    END_CONVERSATION,
+    LOOP,
+    SUREWANTTO,
+    CAPTCHASTATE,
+) = range(8)
